Module crusp::constraints

Modules

arithmetic

handlers

Structs

AllDifferent

Increasing

SumConstraint

Enums

ConstraintState

PropagationState

Traits

Constraint
Trait defining a constraint. The generic type is mandatory for the Constraint, even if the Constraint depends more on its underlying variables. Indeed, the constraints handlers store all the constraint the same, so the constraints are used as trait objects, so they require to have the same type. Having the Handler as a generic parameter to the propagate function will not allow to use Constraint as a trait object.

ConstraintBuilder

Propagator

PropagatorState