Module crusp::variables

Modules

bool_var

domains

handlers

int_var

macros

Structs

ArrayOfRefs
Represents an array of references to Variable.

ArrayOfVars
Represents an array of Variable.

ArrayOfVarsBuilder
Represents an array of Variable builder.

VariableId

Enums

VariableError
Represents an error that occured during variable domain update.

VariableState
Describes the state of a variable after its domain is updated.

Traits

Array
This trait describes an array of variables. There is two types of array: array of variables and array of references to variables. Both types are manipulated with the same trait. When writting constraints over an array of variables, you should use the Array trait instead of the specific types ArrayOfVars or ArrayOfRefs.

ArrayBuilder
Trait used to an array of variable;

Variable
Trait for types that represent decision variables. A decision variable is variable along side with its domain of allowed values. A variable has to be cloneable because the (tree based) searching process is based on cloning.

VariableBuilder
Trait used to build a variable. SpecificVariablesHandler requires to add VariableBuiler or ArrayBuilder in order to assign them one unique id.