Module crusp::variables::domains

Traits

AssignableDomain
Trait that defines variable that can be assigned to a specific value.

FiniteDomain
Trait that defines variables with finite domains. In other words the number of elements of the domain is countable). Every variable should have a finite domain.

FromRangeDomain
Trait that defines variableswhich the domain can be deduced from an interval.

FromValuesDomain
Trait that defines variable which the domain can be deduced from a list of values.

IterableDomain
Trait that definies variable allowing to iter through the elements of its domain.

OrderedDomain
Trait that defines variable which the underlying Type implements the Ord trait (i.e. the underlying type is totally ordered).

OrderedPrunableDomain
Trait that definies variable that allows to remove any values from its domains.

PrunableDomain
Trait that definies variable that allows to remove any values from its domains.