Trait crusp::variables::domains::AssignableDomain

pub trait AssignableDomain: FiniteDomain {
    fn set_value(
        &mut self, 
        value: Self::Type
    ) -> Result<VariableState, VariableError>;
}

Trait that defines variable that can be assigned to a specific value.

Required Methods

fn set_value(
&mut self,
value: Self::Type
) -> Result<VariableState, VariableError>

Change the value of the variable. Returns an error of type VariableError::DomainWipeout if value is not inside the domain, otherwise returns the correct VariableState;

Argument

value - The value to assign.