Trait crusp::variables::domains::OrderedDomain

pub trait OrderedDomain: FiniteDomain where
    Self::Type: Ord + Eq, {
    fn min(&self) -> Option<Self::Type>;

    fn max(&self) -> Option<Self::Type>;

    fn strict_upperbound(
        &mut self, 
        ub: Self::Type
    ) -> Result<VariableState, VariableError>;

    fn weak_upperbound(
        &mut self, 
        ub: Self::Type
    ) -> Result<VariableState, VariableError>;

    fn strict_lowerbound(
        &mut self, 
        lb: Self::Type
    ) -> Result<VariableState, VariableError>;

    fn weak_lowerbound(
        &mut self, 
        lb: Self::Type
    ) -> Result<VariableState, VariableError>;

    fn unchecked_min(&self) -> Self::Type { ... }

    fn unchecked_max(&self) -> Self::Type { ... }

    fn less_than(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError> { ... }

    fn less_or_equal_than(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError> { ... }

    fn greater_than(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError> { ... }

    fn greater_or_equal_than(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError> { ... }

    fn equal_bounds(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError> { ... }
}

Trait that defines variable which the underlying Type implements the Ord trait (i.e. the underlying type is totally ordered).

Required Methods

fn min(&self) -> Option<Self::Type>

Returns the minimal value of the domain.

fn max(&self) -> Option<Self::Type>

Returns the maximal value of the domain.

fn strict_upperbound(
&mut self,
ub: Self::Type
) -> Result<VariableState, VariableError>

Forces the upperbound of the variable to be strictly lesser than ub. Returns an error of type VariableError::DomainWipeout if the new upperbound is lesser than the minimal value of the domain, otherwise returns the correct VariableState.

Parameters

ub - The strict upperbound implied to the domain.

fn weak_upperbound(
&mut self,
ub: Self::Type
) -> Result<VariableState, VariableError>

Forces the upperbound of the variable to be lesser than ub. Returns an error of type VariableError::DomainWipeout if the new upperbound is strictly lesser than the minimal value of the domain, otherwise returns the correct VariableState.

Parameters

ub - The weak upperbound implied to the domain.

fn strict_lowerbound(
&mut self,
lb: Self::Type
) -> Result<VariableState, VariableError>

Forces the lowerbound of the variable to be strictly greater than lb. Returns an error of type VariableError::DomainWipeout if the new lowerbound is greater than the maximal value of the domain, otherwise returns the correct VariableState.

Parameters

ub - The strict lowerbound implied to the domain.

fn weak_lowerbound(
&mut self,
lb: Self::Type
) -> Result<VariableState, VariableError>

Forces the lowerbound of the variable to be greater than lb. Returns an error of type VariableError::DomainWipeout if the new lowerbound is strictly lesser than the maximal value of the domain, otherwise returns the correct VariableState.

Parameters

ub - The weak lowerbound implied to the domain.

Provided Methods

fn unchecked_min(&self) -> Self::Type

fn unchecked_max(&self) -> Self::Type

fn less_than(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the domain of self to satisfies a precedence relation with value. Returns an error of type VariableError::DomainWipeout if the minimal value of self is greater or equal to the maximal value of value, otherwise returns the correct VariableState.

Parameters

value - The variable to compare to.

fn less_or_equal_than(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the domain of self to satisfies a weak precedence relation with value. Returns an error of type VariableError::DomainWipeout if the minimal value of self is greater to the maximal value of value, otherwise returns the correct VariableState.

Parameters

value - The variable to compare to.

fn greater_than(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the domain of value to satisfies a strict precedence relation with self. Returns an error of type VariableError::DomainWipeout if the minimal value of value is greater or equal to the maximal value of self, otherwise returns the correct VariableState.

Parameters

value - The variable to compare to.

fn greater_or_equal_than(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the domain of value to satisfies a weak precedence relation with self. Returns an error of type VariableError::DomainWipeout if the minimal value of value is greater to the maximal value of self, otherwise returns the correct VariableState.

Parameters

value - The variable to compare to.

fn equal_bounds(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the domains of two variables two have the same bounds (the does not imply to have the same domain). Returns an error of type VariableError::DomainWipeout if the two variables can't have the same bounds (i.e. no common value), otherwise returns the correct VariableState.

Parameters

value - The variable to compare to.