Trait crusp::variables::domains::OrderedPrunableDomain

pub trait OrderedPrunableDomain: PrunableDomain + OrderedDomain where
    Self::Type: Eq + Ord, {
    fn in_sorted_values<Values: Iterator<Item = Self::Type>>(
        &mut self, 
        values: Values
    ) -> Result<VariableState, VariableError>
    where
        Values: IntoIterator<Item = Self::Type>;
}

Trait that definies variable that allows to remove any values from its domains.

Required Methods

fn in_sorted_values<Values: Iterator<Item = Self::Type>>(
&mut self,
values: Values
) -> Result<VariableState, VariableError> where
Values: IntoIterator<Item = Self::Type>,

Forces the domain of the variables to be in the values past has parameter.

Parameters

value - The variable to compare to.