Trait crusp::variables::domains::PrunableDomain

pub trait PrunableDomain: FiniteDomain where
    Self::Type: Eq, {
    fn equal(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError>;

    fn not_equal(
        &mut self, 
        value: &mut Self
    ) -> Result<(VariableState, VariableState), VariableError>;

    fn in_values<Values>(
        &mut self, 
        values: Values
    ) -> Result<VariableState, VariableError>
    where
        Values: IntoIterator<Item = Self::Type>;

    fn remove_value(
        &mut self, 
        value: Self::Type
    ) -> Result<VariableState, VariableError>;

    fn remove_if<Predicate>(
        &mut self, 
        pred: Predicate
    ) -> Result<VariableState, VariableError>
    where
        Predicate: FnMut(&Self::Type) -> bool;

    fn retains_if<Predicate>(
        &mut self, 
        pred: Predicate
    ) -> Result<VariableState, VariableError>
    where
        Predicate: FnMut(&Self::Type) -> bool;
}

Trait that definies variable that allows to remove any values from its domains.

Required Methods

fn equal(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the domain of two variables to be equal.

Parameters

value - The variable to compare to.

fn not_equal(
&mut self,
value: &mut Self
) -> Result<(VariableState, VariableState), VariableError>

Forces the value of two varaibles to be distinct.

Parameters

value - The variable to compare to.

fn in_values<Values>(
&mut self,
values: Values
) -> Result<VariableState, VariableError> where
Values: IntoIterator<Item = Self::Type>,

Forces the domain of the variables to be in the values past has parameter.

Parameters

value - The variable to compare to.

fn remove_value(
&mut self,
value: Self::Type
) -> Result<VariableState, VariableError>

Remove the value from the domain of a variable.

Parameters

value - The variable to compare to.

fn remove_if<Predicate>(
&mut self,
pred: Predicate
) -> Result<VariableState, VariableError> where
Predicate: FnMut(&Self::Type) -> bool,

Remove the values of the domain that satisfies the predicate.

Parameters

value - The variable to compare to.

fn retains_if<Predicate>(
&mut self,
pred: Predicate
) -> Result<VariableState, VariableError> where
Predicate: FnMut(&Self::Type) -> bool,

Remove the values of the domain that does not satisfy the predicate.

Parameters

value - The variable to compare to.