Trait crusp::variables::Array

pub trait Array {
    type Variable: Variable;
    fn get_mut(&mut self, position: usize) -> Option<&mut Self::Variable>;

    fn get(&self, position: usize) -> Option<&Self::Variable>;

    fn get_unchecked_mut(&mut self, position: usize) -> &mut Self::Variable;

    fn get_unchecked(&self, position: usize) -> &Self::Variable;

    fn iter<'array>(
        &'array self
    ) -> Box<Iterator<Item = &Self::Variable> + 'array>;

    fn iter_mut<'array>(
        &'array mut self
    ) -> Box<Iterator<Item = &mut Self::Variable> + 'array>;

    fn len(&self) -> usize;

    fn iter_ids<'a>(&'a self) -> Box<Iterator<Item = VariableId> + 'a> { ... }
}

This trait describes an array of variables. There is two types of array: array of variables and array of references to variables. Both types are manipulated with the same trait. When writting constraints over an array of variables, you should use the Array trait instead of the specific types ArrayOfVars or ArrayOfRefs.

Associated Types

type Variable: Variable

Required Methods

fn get_mut(&mut self, position: usize) -> Option<&mut Self::Variable>

Returns a mutable reference to the variable at that position or None if out of bounds.

fn get(&self, position: usize) -> Option<&Self::Variable>

Returns a reference to the variable at that position or None if out of bounds.

fn get_unchecked_mut(&mut self, position: usize) -> &mut Self::Variable

Returns a mutable reference to the variable at that position without doing bounds check.

fn get_unchecked(&self, position: usize) -> &Self::Variable

Returns a reference to the variable at that position without doing bounds check.

fn iter<'array>(&'array self) -> Box<Iterator<Item = &Self::Variable> + 'array>

Returns an iterator over the variables.

fn iter_mut<'array>(
&'array mut self
) -> Box<Iterator<Item = &mut Self::Variable> + 'array>

Returns an iterator that allows modifying each variable.

fn len(&self) -> usize

Returns the number of variables.

Provided Methods

fn iter_ids<'a>(&'a self) -> Box<Iterator<Item = VariableId> + 'a>