Trait crusp::variables::Variable

pub trait Variable: Clone {
    type Type: Clone;
    fn is_affected(&self) -> bool;

    fn value(&self) -> Option<Self::Type>;

    fn id(&self) -> VariableId;

    fn iter_ids(&self) -> Box<Iterator<Item = VariableId>> { ... }
}

Trait for types that represent decision variables. A decision variable is variable along side with its domain of allowed values. A variable has to be cloneable because the (tree based) searching process is based on cloning.

Associated Types

type Type: Clone

The underlying type holded by the Variable.

Required Methods

fn is_affected(&self) -> bool

Returns if the variable is affected. A variable is affected if and only if its a domain is a singleton.

fn value(&self) -> Option<Self::Type>

Returns the value of the variable or None if the variable is not affected.

fn id(&self) -> VariableId

Returns the state of the variable without reinitialising it. The state of a variable describes if and how the domain of the variable has been updated.

Provided Methods

fn iter_ids(&self) -> Box<Iterator<Item = VariableId>>